// src/app/page/master/pegawai-sk-tahapan-rekruitment/pegawai-sk-tahapan-rekruitment.interface.ts
import {EventEmitter} from '@angular/core';

export interface PegawaiSKTahapanRekruitment {
	kdProfile?: string;
	kdAgama?: string;
	namaAgama?: string;
	reportDisplay?: string;
	kodeExternal?: string;
	namaExternal?: string;
	statusEnabled?: boolean;
	noRec?: string;
	id?: any;
	kode?: any;
	version?: any;
	label?: any;
	value?: any;
	nama?: any;
	kodeEx?: any;
	namaEx?: any;

 }